requirements.txt
tqdm
opencv-python
scipy
numpy
easydict
PyYAML
scikit-learn